#11 move liquibase plugin and resources to app for now
