+++ data-sync-app/pom.xml
@@ -15,22 +15,18 @@
         <dependency>
             <groupId>com.github.zybercik00</groupId>
             <artifactId>data-sync-repository</artifactId>
-            <version>1.0-SNAPSHOT</version>
         </dependency>
         <dependency>
             <groupId>com.github.zybercik00</groupId>
             <artifactId>data-sync-service</artifactId>
-            <version>1.0-SNAPSHOT</version>
         </dependency>
         <dependency>
             <groupId>com.github.zybercik00</groupId>
             <artifactId>data-sync-excel</artifactId>
-            <version>1.0-SNAPSHOT</version>
         </dependency>
         <dependency>
             <groupId>com.github.zybercik00</groupId>
             <artifactId>data-sync-config</artifactId>
-            <version>1.0-SNAPSHOT</version>
         </dependency>
 
         <dependency>
@@ -61,4 +57,29 @@
         </dependency>
     </dependencies>
 
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>org.liquibase</groupId>
+                <artifactId>liquibase-maven-plugin</artifactId>
+                <executions>
+                    <execution>
+                        <id>generateChangeLog</id>
+                        <goals>
+                            <goal>generateChangeLog</goal>
+                        </goals>
+                        <phase>process-resources</phase>
+                        <configuration>
+                            <driver>org.h2.Driver</driver>
+                            <url>jdbc:h2:file:./data/DataSyncExcel</url>
+                            <username>sa</username>
+                            <password>password</password>
+                            <outputChangeLogFile>src/main/resources/liquibase-outputChangeLog.xml</outputChangeLogFile>
+                            <overwriteOutputFile>true</overwriteOutputFile>
+                        </configuration>
+                    </execution>
+                </executions>
+            </plugin>
+        </plugins>
+    </build>
 </project>
