+++ data-sync-init/src/main/java/com/github/zybercik00/DataInitializerService.java
@@ -1,8 +1,8 @@
 package com.github.zybercik00;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.github.zybercik00.repository.mappingattribute.MappingRepo;
-import com.github.zybercik00.repository.mappingattribute.AttributeRepo;
+import com.github.zybercik00.repository.metadata.MappingRepo;
+import com.github.zybercik00.repository.metadata.AttributeRepo;
 import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
