+++ data-sync-repository/src/main/java/com/github/zybercik00/repository/metadata/AttributeRepo.java
@@ -1,6 +1,6 @@
-package com.github.zybercik00.repository.mappingattribute;
+package com.github.zybercik00.repository.metadata;
 
-import com.github.zybercik00.domain.metadata.AttributeEntity;
+import com.github.zybercik00.metadata.AttributeEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AttributeRepo extends JpaRepository<AttributeEntity, String> {
