+++ data-sync-repository/src/main/java/com/github/zybercik00/repository/metadata/MappingRepo.java
@@ -1,4 +1,4 @@
-package com.github.zybercik00.repository.mappingattribute;
+package com.github.zybercik00.repository.metadata;
 
 import com.github.zybercik00.metadata.MappingEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
